+++ server/src-lib/Hasura/RQL/DDL/RemoteRelationship/Validate.hs
@@ -363,13 +363,21 @@ validateType permittedVariables value expectedGType types =
         Just fieldInfo ->
           bindValidation
             (columnInfoToNamedType fieldInfo)
-            (\actualNamedType -> assertType (G.toGT actualNamedType) expectedGType)
-    G.VInt {} -> assertType (G.toGT $ mkScalarTy PGInteger) expectedGType
-    G.VFloat {} -> assertType (G.toGT $ mkScalarTy PGFloat) expectedGType
-    G.VBoolean {} -> assertType (G.toGT $ mkScalarTy PGBoolean) expectedGType
-    G.VNull -> Failure (pure NullNotAllowedHere)
-    G.VString {} -> assertType (G.toGT $ mkScalarTy PGText) expectedGType
-    G.VEnum _ -> Failure (pure UnsupportedEnum)
+            (\actualNamedType -> isTypeCoercible (G.toGT actualNamedType) expectedGType)
+    G.VInt {} -> do
+      let intScalarGType = G.toGT $ mkScalarTy PGInteger
+      isTypeCoercible intScalarGType expectedGType
+    G.VFloat {} -> do
+      let floatScalarGType = G.toGT $ mkScalarTy PGFloat
+      isTypeCoercible floatScalarGType expectedGType
+    G.VBoolean {} -> do
+      let boolScalarGType = G.toGT $ mkScalarTy PGBoolean
+      isTypeCoercible boolScalarGType expectedGType
+    G.VNull -> Failure $ pure NullNotAllowedHere
+    G.VString {} -> do
+      let stringScalarGType = G.toGT $ mkScalarTy PGText
+      isTypeCoercible stringScalarGType expectedGType
+    G.VEnum _ -> Failure $ pure UnsupportedEnum
     G.VList (G.unListValue -> values) -> do
       case values of
         []  -> pure ()
@@ -405,19 +413,26 @@ validateType permittedVariables value expectedGType types =
                         (G.toGT $ G.NamedType name)
                         "not an input object type"))
 
-assertType :: G.GType -> G.GType -> Validation (NonEmpty ValidationError) ()
-assertType actualType expectedType = do
-  -- check if both are list types or both are named types
-  (when
-     (isListType' actualType /= isListType' expectedType)
-     (Failure (pure $ ExpectedTypeButGot expectedType actualType)))
-  -- if list type then check over unwrapped type, else check base types
-  if isListType' actualType
-    then assertType (unwrapTy actualType) (unwrapTy expectedType)
-    else (when
-            (getBaseTy actualType /= getBaseTy expectedType)
-            (Failure (pure $ ExpectedTypeButGot expectedType actualType)))
-  pure ()
+isTypeCoercible
+  :: G.GType
+  -> G.GType
+  -> Validation (NonEmpty ValidationError) ()
+isTypeCoercible actualType expectedType =
+  -- The GraphQL spec says that, a singleton type can be coerced into  an array
+  -- type. Which means that if the 'actualType' is a singleton type, like
+  -- 'Int' we should be able to join this with a remote node, which expects an
+  -- input argument of type '[Int]'
+  -- http://spec.graphql.org/June2018/#sec-Type-System.List
+  let (actualBaseType, actualNestingLevel) = getBaseTyWithNestedLevelsCount actualType
+      (expectedBaseType, expectedNestingLevel) = getBaseTyWithNestedLevelsCount expectedType
+  in
+  if | actualBaseType /= expectedBaseType -> raiseValidationError
+       -- we cannot coerce two types with different nesting levels,
+       -- for example, we cannot coerce [Int] to [[Int]]
+     | (actualNestingLevel == expectedNestingLevel || actualNestingLevel == 0) -> pure ()
+     | otherwise -> raiseValidationError
+     where
+       raiseValidationError = Failure $ pure $ ExpectedTypeButGot expectedType actualType
 
 assertListType :: G.GType -> Validation (NonEmpty ValidationError) ()
 assertListType actualType =